+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Added config option to enable midair border spawning (great for skyblock or similar), by default it is disabled
 - Added config options to disable biome and block checks for spawning birds
 - Added new command "spawn" to spawn a bird of a specific type at the player's location
+- Changed Chunk Load Spawning to spawn only half of max birds per species, instead of the full amount
 
 ---